Only move vim_related if it doesn't exist

+++ main.sh
@@ -15,6 +15,13 @@ else
   echo "ahoym/dotfiles already [include]d in ~/.gitconfig. Skipping [include-git]."
 fi
 
-# Copy vim related files/dirs to root
-cp -R ./vim_related/ ~
+# [include-vim]
+VIM_COPY_SIGIL="File copied from ahoym/vim_related"
+# Copies vim related files/dirs if not existing
+if [ ! -f ~/.vimrc ] || ! grep -q "$VIM_COPY_SIGIL" ~/.vimrc; then
+  echo "Copying vim related files/dirs to root"
+  cp -R ./vim_related/ ~
+else
+  echo "ahoym/vim_related already exists in ~/.vimrc. Skipping [include-vim]."
+fi
